fix: move config pkg to operator
