+++ operator/batch/internal/syncerfactory_mock.go
@@ -12,8 +12,8 @@ package internal
 import (
 	reflect "reflect"
 
+	config "github.com/equinor/radix-operator/operator/config"
 	batch "github.com/equinor/radix-operator/pkg/apis/batch"
-	config "github.com/equinor/radix-operator/pkg/apis/config"
 	kube "github.com/equinor/radix-operator/pkg/apis/kube"
 	v1 "github.com/equinor/radix-operator/pkg/apis/radix/v1"
 	versioned "github.com/equinor/radix-operator/pkg/client/clientset/versioned"
