+++ operator/config/containerregistry_config.go
