+++ operator/config/gateway_config.go
