+++ operator/dnsalias/handler.go
@@ -4,8 +4,8 @@ import (
 	"context"
 
 	"github.com/equinor/radix-operator/operator/common"
+	"github.com/equinor/radix-operator/operator/config"
 	"github.com/equinor/radix-operator/operator/dnsalias/internal"
-	"github.com/equinor/radix-operator/pkg/apis/config"
 	"github.com/equinor/radix-operator/pkg/apis/defaults"
 	"github.com/equinor/radix-operator/pkg/apis/dnsalias"
 	"github.com/equinor/radix-operator/pkg/apis/ingress"
