+++ operator/dnsalias/handler_test.go
@@ -6,9 +6,9 @@ import (
 	"testing"
 
 	"github.com/equinor/radix-operator/operator/common"
+	"github.com/equinor/radix-operator/operator/config"
 	"github.com/equinor/radix-operator/operator/dnsalias"
 	"github.com/equinor/radix-operator/operator/dnsalias/internal"
-	"github.com/equinor/radix-operator/pkg/apis/config"
 	"github.com/equinor/radix-operator/pkg/apis/defaults"
 	dnsaliasapi "github.com/equinor/radix-operator/pkg/apis/dnsalias"
 	"github.com/equinor/radix-operator/pkg/apis/ingress"
