+++ operator/environment/handler.go
@@ -3,7 +3,7 @@ package environment
 import (
 	"context"
 
-	"github.com/equinor/radix-operator/pkg/apis/config"
+	"github.com/equinor/radix-operator/operator/config"
 	"github.com/equinor/radix-operator/pkg/apis/networkpolicy"
 	"github.com/rs/zerolog/log"
