+++ operator/main.go
@@ -19,6 +19,7 @@ import (
 	"github.com/equinor/radix-operator/operator/application"
 	"github.com/equinor/radix-operator/operator/batch"
 	"github.com/equinor/radix-operator/operator/common"
+	apiconfig "github.com/equinor/radix-operator/operator/config"
 	"github.com/equinor/radix-operator/operator/deployment"
 	"github.com/equinor/radix-operator/operator/dnsalias"
 	"github.com/equinor/radix-operator/operator/environment"
@@ -26,7 +27,6 @@ import (
 	"github.com/equinor/radix-operator/operator/registration"
 	"github.com/equinor/radix-operator/operator/scheduler"
 	"github.com/equinor/radix-operator/operator/scheduler/tasks"
-	apiconfig "github.com/equinor/radix-operator/pkg/apis/config"
 	"github.com/equinor/radix-operator/pkg/apis/defaults"
 	"github.com/equinor/radix-operator/pkg/apis/event"
 	"github.com/equinor/radix-operator/pkg/apis/ingress"
